requirements.txt
python-dotenv
requests
selenium
pillow
python-telegram-bot
pyautogui
webdriver-manager